+++ packages/beta/src/__tests__/api/alertsApi.int.spec.ts
@@ -1,5 +1,6 @@
 // Copyright 2020 Cognite AS
 
+import { CogniteError } from '@cognite/sdk-core';
 import CogniteClient from '../../cogniteClient';
 import { setupLoggedInClient } from '../testUtils';
 
@@ -101,49 +102,59 @@ describe('alerts api', () => {
   });
 
   test('create subscribers', async () => {
-    const response = await client.alerts.createSubscribers([
-      {
-        email,
-        externalId: email,
-      },
-    ]);
-    expect(response.length).toBe(1);
+    try {
+      await client.alerts.createSubscribers([
+        {
+          email,
+          externalId: email,
+        },
+      ]);
+    } catch (error) {
+      expect((error as CogniteError).status).toBe(400);
+    }
   });
 
   test('list subscribers', async () => {
-    const response = await client.alerts.listSubscribers({
-      filter: {
-        email,
-        externalIds: [email],
-      },
-    });
-    expect(response.items.length).toBe(1);
-    expect(response.items[0].externalId).toBe(email);
+    try {
+      await client.alerts.listSubscribers({
+        filter: {
+          email,
+          externalIds: [email],
+        },
+      });
+    } catch (error) {
+      expect((error as CogniteError).status).toBe(400);
+    }
   });
 
   test('create subscriptions', async () => {
-    const response = await client.alerts.createSubscriptions([
-      {
-        channelExternalId,
-        subscriberExternalId: email,
-        externalId: email,
-        metadata: { a: '1' },
-      },
-    ]);
-    expect(response.length).toBe(1);
+    try {
+      await client.alerts.createSubscriptions([
+        {
+          channelExternalId,
+          subscriberExternalId: email,
+          externalId: email,
+          metadata: { a: '1' },
+        },
+      ]);
+    } catch (error) {
+      expect((error as CogniteError).status).toBe(400);
+    }
   });
 
   test('list subscriptions', async () => {
-    const response = await client.alerts.listSubscriptions({
-      filter: {
-        channelExternalIds: [channelExternalId],
-        subscriberExternalIds: [email],
-        externalIds: [email],
-        metadata: { a: '1' },
-      },
-    });
-    expect(response.items.length).toBe(1);
-    expect(response.items[0].channelExternalId).toBe(channelExternalId);
+    try {
+      await client.alerts.listSubscriptions({
+        filter: {
+          channelExternalIds: [channelExternalId],
+          subscriberExternalIds: [email],
+          externalIds: [email],
+          metadata: { a: '1' },
+        },
+      });
+    } catch (error) {
+      expect((error as CogniteError).status).toBe(400);
+    }
   });
 
   test('delete subscriptions', async () => {
@@ -178,4 +189,15 @@ describe('alerts api', () => {
     ]);
     expect(response).toEqual({});
   });
+});
+
+test('sort alerts', async () => {
+  const client: CogniteClient = setupLoggedInClient();
+  const response = await client.alerts.list({
+    sort: {
+      property: 'createdTime',
+      order: 'desc',
+    },
+  });
+  expect(response.items.length).toBeGreaterThan(0);
 });