+++ packages/beta/src/cogniteClient.ts
@@ -9,6 +9,7 @@ import { version } from '../package.json';
 import { AlertsAPI } from './api/alerts/alertsApi';
 import { FilesAPI } from './api/files/filesApi';
 import { MonitoringTasksAPI } from './api/monitoringTasks/monitoringTasksApi';
+import { DataPointsAPI } from './api/dataPoints/dataPointsApi';
 
 class CogniteClientCleaned extends CogniteClientStable {
   // Remove type restrictions
@@ -51,6 +52,12 @@ export default class CogniteClient extends CogniteClientCleaned {
     return accessApi(this.monitoringTasksApi);
   }
 
+  public get datapoints() {
+    return accessApi(this.dataPointsApi);
+  }
+
+  protected dataPointsApi?: DataPointsAPI;
+
   protected get version() {
     return `${version}-beta`;
   }
@@ -66,5 +73,6 @@ export default class CogniteClient extends CogniteClientCleaned {
       MonitoringTasksAPI,
       'monitoringtasks'
     );
+    this.dataPointsApi = this.apiFactory(DataPointsAPI, 'timeseries/data');
   }
 }