+++ packages/beta/src/index.ts
@@ -4,3 +4,20 @@ export * from '@cognite/sdk';
 export { FilesMultipartUploadSessionAPI } from './api/files/filesMultipartUploadSessionApi';
 export { default as CogniteClient } from './cogniteClient';
 export * from './types';
+export {
+  Aggregate,
+  DatapointAggregate,
+  DatapointAggregates,
+  Datapoints,
+  DatapointsMonthlyGranularityMultiQuery,
+  DatapointsMultiQuery,
+  DatapointsMultiQueryBase,
+  DatapointsQuery,
+  DatapointsQueryExternalId,
+  DatapointsQueryId,
+  DatapointsQueryProperties,
+  DoubleDatapoint,
+  DoubleDatapoints,
+  StringDatapoint,
+  StringDatapoints,
+} from './types';